Move all the policies registrar test cases in the same module.
