+++ st2common/tests/unit/test_policies.py
@@ -13,19 +13,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
-
-import jsonschema
-
-import st2tests
-import st2common
-from st2common.bootstrap.policiesregistrar import PolicyRegistrar
-from st2common.bootstrap.policiesregistrar import register_policy_types, register_policies
 from st2common.persistence.policy import PolicyType, Policy
 from st2common.policies import ResourcePolicyApplicator, get_driver
-from st2tests import DbTestCase, fixturesloader
+from st2tests import DbTestCase
 from st2tests.fixturesloader import FixturesLoader
-from st2tests.fixturesloader import get_fixtures_packs_base_path
+
+__all__ = [
+    'PolicyTestCase'
+]
 
 PACK = 'generic'
 TEST_FIXTURES = {
@@ -46,11 +41,11 @@ TEST_FIXTURES = {
 }
 
 
-class PolicyTest(DbTestCase):
+class PolicyTestCase(DbTestCase):
 
     @classmethod
     def setUpClass(cls):
-        super(PolicyTest, cls).setUpClass()
+        super(PolicyTestCase, cls).setUpClass()
 
         loader = FixturesLoader()
         loader.save_fixtures_to_db(fixtures_pack=PACK,
@@ -75,59 +70,3 @@ class PolicyTest(DbTestCase):
         self.assertEqual(policy._policy_type, policy_db.policy_type)
         self.assertTrue(hasattr(policy, 'threshold'))
         self.assertEqual(policy.threshold, 3)
-
-
-class PolicyBootstrapTest(DbTestCase):
-
-    @classmethod
-    def setUpClass(cls):
-        super(PolicyBootstrapTest, cls).setUpClass()
-
-        # Register common policy types
-        register_policy_types(st2common)
-
-    def test_register_policy_types(self):
-        self.assertEqual(register_policy_types(st2tests), 2)
-
-        type1 = PolicyType.get_by_ref('action.concurrency')
-        self.assertEqual(type1.name, 'concurrency')
-        self.assertEqual(type1.resource_type, 'action')
-
-        type2 = PolicyType.get_by_ref('action.mock_policy_error')
-        self.assertEqual(type2.name, 'mock_policy_error')
-        self.assertEqual(type2.resource_type, 'action')
-
-    def test_register_policies(self):
-        # Note: Only one policy should be registered since second one fails validation
-        pack_dir = os.path.join(fixturesloader.get_fixtures_packs_base_path(), 'dummy_pack_1')
-        self.assertEqual(register_policies(pack_dir=pack_dir), 1)
-
-        p1 = Policy.get_by_ref('dummy_pack_1.test_policy_1')
-        self.assertEqual(p1.name, 'test_policy_1')
-        self.assertEqual(p1.pack, 'dummy_pack_1')
-        self.assertEqual(p1.resource_ref, 'dummy_pack_1.local')
-        self.assertEqual(p1.policy_type, 'action.concurrency')
-
-        p2 = Policy.get_by_ref('dummy_pack_1.test_policy_2')
-        self.assertEqual(p2, None)
-
-    def test_register_policy_invalid_policy_type_references(self):
-        # Policy references an invalid (inexistent) policy type
-        registrar = PolicyRegistrar()
-        policy_path = os.path.join(get_fixtures_packs_base_path(),
-                                   'dummy_pack_1/policies/policy_2.yaml')
-
-        expected_msg = 'Referenced policy_type "action.mock_policy_error" doesnt exist'
-        self.assertRaisesRegexp(ValueError, expected_msg, registrar._register_policy,
-                                pack='dummy_pack_1', policy=policy_path)
-
-    def test_make_sure_policy_parameters_are_validated_during_register(self):
-        # Policy where specified parameters fail schema validation
-        registrar = PolicyRegistrar()
-        policy_path = os.path.join(get_fixtures_packs_base_path(),
-                                   'dummy_pack_1/policies/policy_3.yaml')
-
-        expected_msg = '100 is greater than the maximum of 5'
-        self.assertRaisesRegexp(jsonschema.ValidationError, expected_msg,
-                                registrar._register_policy, pack='dummy_pack_1',
-                                policy=policy_path)
