+++ st2tests/st2tests/fixtures/packs/dummy_pack_2/policies/policy_3.yaml
@@ -1,10 +1,11 @@
 ---
+# Policy where parameters fail schema validation
 name: test_policy_3
-pack: dummy_pack_1
+pack: dummy_pack_2
 description: Test policy.
 enabled: true
 resource_ref: dummy_pack_1.local
 policy_type: action.retry
 parameters:
     retry_on: timeout
-    max_retry_count: 5
+    max_retry_count: 100
